+++ sdk/datafactory/Microsoft.Azure.Management.DataFactory/tests/JsonSamples/LinkedServiceJsonSamples.cs
@@ -27,7 +27,12 @@ namespace DataFactory.Tests.JsonSamples
              value: ""mykey""
           },
           tenant: ""72f988bf-86f1-41af-91ab-2d7cd011db47"",
-          azureCloudType: ""AzurePublic""
+          azureCloudType: ""AzurePublic"",
+          servicePrincipalCredentialType: ""ServicePrincipalKey"",
+          servicePrincipalCredential: {
+             type: ""SecureString"",
+             value: ""mykey""
+          },
         }
     }
 }";